+++ sakura_core/config/app_constants.h
@@ -52,16 +52,16 @@
 // -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
 
 // Feb. 18, 2003 genta 最大値の定数化と値変更
-const int LINESPACE_MAX = 128;
-const int COLUMNSPACE_MAX = 64;
+const size_t LINESPACE_MAX = 128;
+const size_t COLUMNSPACE_MAX = 64;
 
 // Aug. 14, 2005 genta 定数定義追加
 // 2007.09.07 kobake 定数名変更: MAXLINESIZE→MAXLINEKETAS
 // 2007.09.07 kobake 定数名変更: MINLINESIZE→MINLINEKETAS
-const int MAXLINEKETAS		= 10240;	// 1行の桁数の最大値
-const int MINLINEKETAS		= 10;		// 1行の桁数の最小値
+const size_t MAXLINEKETAS		= 10240;	// 1行の桁数の最大値
+const size_t MINLINEKETAS		= 10;		// 1行の桁数の最小値
 
 // 2014.08.02 定数定義追加 katze
-const int LINENUMWIDTH_MIN = 2;
-const int LINENUMWIDTH_MAX = 11;
+const size_t LINENUMWIDTH_MIN = 2;
+const size_t LINENUMWIDTH_MAX = 11;
